src/heatmap-mapbox/requirements.txt
flask
pandas
cdsapi
netCDF4
pandas
numpy
xarray
matplotlib
cartopy
setuptools
python-dotenv
Flask-Caching

